requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-appconf==1.0.4
django-heroku==0.3.1
django-imagekit==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
drf-writable-nested==0.6.0
gunicorn==20.0.4
pilkit==2.0
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.2.0
